Fix outbound.completed implementation
Prior to this commit, the url generated by client.outbound.completed()
included a url-encoded tuple where a comma-separated list was called
for, resulting in a HTTP 400 when called.

The method now pre-processes the arguments, generating the string that
is expected to be in the url.

Local testing shows that it now functions, and the relevant test has
been updated to match.